@@ -4,16 +4,18 @@
   "private": true,
   "dependencies": {
     "@daml.js/daml-ui-template-0.0.1": "file:../daml.js/daml-ui-template-0.0.1",
-    "@daml/ledger": "1.15.0",
-    "@daml/react": "1.15.0",
-    "@daml/types": "1.15.0",
-    "@material-ui/core": "^4.3.0",
-    "@material-ui/icons": "^4.2.1",
-    "@material-ui/styles": "^4.3.0",
+    "@daml/ledger": "1.18.1",
+    "@daml/react": "1.18.1",
+    "@daml/types": "1.18.1",
+    "@material-ui/core": "^4.12.3",
+    "@material-ui/icons": "^4.11.2",
+    "@material-ui/styles": "^4.11.4",
+    "@mui/icons-material": "^5.5.1",
+    "@mui/material": "^5.5.2",
     "jsonwebtoken": "^8.5.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-router-dom": "^5.0.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.2",
     "tinycolor2": "^1.4.1"
   },
   "resolutions": {
@@ -47,11 +49,10 @@
   "proxy": "http://localhost:7575",
   "devDependencies": {
     "@types/jsonwebtoken": "^8.5.0",
-    "@types/react": "^16.9.36",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-router-dom": "^5.1.5",
+    "@types/react": "^17.0.2",
+    "@types/react-dom": "^17.0.2",
     "@types/tinycolor2": "^1.4.2",
-    "react-scripts": "^3.4.1",
-    "typescript": "^4.2.3"
+    "react-scripts": "^2.1.3",
+    "typescript": "^4.6.2"
   }
 }
