@@ -1,5 +1,5 @@
 import React from "react";
-import { Route, Switch, withRouter } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 import DamlLedger from "@daml/react";
 import Header from "../Header/Header";
 import Sidebar from "../Sidebar/Sidebar";
@@ -7,10 +7,12 @@ import Report from "../../pages/report/Report";
 import { useUserState } from "../../context/UserContext";
 import { wsBaseUrl, httpBaseUrl } from "../../config";
 import useStyles from "./styles";
+import Login from "../../pages/login/Login";
 
 const Layout = () => {
   const classes = useStyles();
-  const user = useUserState();
+  // const user = useUserState();
+  const [user, setUser] = React.useState<any>(useUserState());
 
   if(!user.isAuthenticated){
     return null;
@@ -23,15 +25,16 @@ const Layout = () => {
               <Sidebar />
               <div className={classes.content}>
                 <div className={classes.fakeToolbar} />
-                <Switch>
-                  <Route path="/app/report" component={Report} />
-                </Switch>
+                <Routes>
+                  <Route path="/app/report" element={Report} />
+                </Routes>
               </div>
             </>
         </div>
       </DamlLedger>
+
     );
   }
 }
 
-export default withRouter(Layout);
+export default Layout;
